+++ framework/db-core/src/main/java/is/codion/framework/db/DefaultUpdate.java
@@ -22,7 +22,7 @@ final class DefaultUpdate implements Update, Serializable {
   private final Condition where;
   private final Map<Column<?>, Object> values;
 
-  private DefaultUpdate(DefaultUpdate.DefaultBuilder builder) {
+  private DefaultUpdate(DefaultBuilder builder) {
     this.where = builder.where;
     this.values = unmodifiableMap(builder.values);
   }
@@ -84,6 +84,10 @@ final class DefaultUpdate implements Update, Serializable {
 
     @Override
     public Update build() {
+      if (values.isEmpty()) {
+        throw new IllegalStateException("No values provided for update");
+      }
+
       return new DefaultUpdate(this);
     }
   }